+++ tests/drizzle-orm/index.test.mjs
@@ -2,7 +2,7 @@ import { sql } from 'drizzle-orm';
 import { drizzle } from 'drizzle-orm/postgres-js';
 import { pgTable, serial } from 'drizzle-orm/pg-core';
 import postgres from 'postgres';
-import { cosineDistance, l2Distance, maxInnerProduct, l1Distance, vector } from 'pgvector/drizzle-orm';
+import { cosineDistance, l2Distance, maxInnerProduct, l1Distance, vector, halfvec } from 'pgvector/drizzle-orm';
 
 test('example', async () => {
   const client = postgres({database: 'pgvector_node_test', onnotice: function() {}});
@@ -10,11 +10,12 @@ test('example', async () => {
 
   await client`CREATE EXTENSION IF NOT EXISTS vector`;
   await client`DROP TABLE IF EXISTS drizzle_items`;
-  await client`CREATE TABLE drizzle_items (id serial PRIMARY KEY, embedding vector(3))`;
+  await client`CREATE TABLE drizzle_items (id serial PRIMARY KEY, embedding vector(3), half_embedding halfvec(3))`;
 
   const items = pgTable('drizzle_items', {
     id: serial('id').primaryKey(),
-    embedding: vector('embedding', {dimensions: 3})
+    embedding: vector('embedding', {dimensions: 3}),
+    halfEmbedding: halfvec('half_embedding', {dimensions: 3})
   });
 
   const newItems = [