+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added support for `halfvec` and `sparsevec` types to Knex.js
 - Added support for `halfvec` type to MikroORM
+- Added support for `halfvec` type to Drizzle ORM
 - Added `l1Distance`, `hammingDistance`, and `jaccardDistance` functions for Knex.js
 - Added `l1Distance`, `hammingDistance`, and `jaccardDistance` functions for Objection.js
 - Added `l1Distance`, `hammingDistance`, and `jaccardDistance` functions for Kysely